Allow to pass custom GET parameters.
Stripping away all GET parameters is a good way to
make them usable to differentiate between the PDF
and HTML view. But sometimes the application logic
behind the view also needs some GET parameters to
function properly or needs to know that it's beeing
accessed by the printer-request.

This allows to manually pass those parameters as a
dictionary to the request_to_pdf method.